requirements.txt
streamlit
pandas
Pillow
fpdf
